+++ pubsub/runtime/src/main/java/io/quarkiverse/googlecloudservices/pubsub/QuarkusPubSub.java
@@ -7,8 +7,6 @@ import java.util.stream.StreamSupport;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import jakarta.enterprise.context.ApplicationScoped;
-import jakarta.enterprise.inject.Disposes;
-import jakarta.enterprise.inject.Produces;
 import jakarta.enterprise.inject.Instance;
 import jakarta.inject.Inject;
 
@@ -103,21 +101,6 @@ public class QuarkusPubSub {
     }
 
     /**
-     * Makes the subscription admin client available as CDI bean
-     */
-    @Produces
-    public SubscriptionAdminClient subscriptionAdminClient() throws IOException {
-        return SubscriptionAdminClient.create(subscriptionAdminSettings());
-    }
-
-    /**
-     * CDI Dispose method for {@link #subscriptionAdminClient()}. Shouldn't be called directly
-     */
-    public void shutdownSubscriptionAdminClient(@Disposes SubscriptionAdminClient subscriptionAdminClient) {
-        subscriptionAdminClient.close();
-    }
-
-    /**
      * Creates a PubSub TopicAdminSettings using the configured project ID.
      */
     public TopicAdminSettings topicAdminSettings() throws IOException {
@@ -125,21 +108,6 @@ public class QuarkusPubSub {
                 .setCredentialsProvider(credentialsProvider());
         channelProvider.ifPresent(builder::setTransportChannelProvider);
         return builder.build();
-    }
-
-    /**
-     * Makes the topic admin client available as a CDI bean
-     */
-    @Produces
-    public TopicAdminClient topicAdminClient() throws IOException {
-        return TopicAdminClient.create(topicAdminSettings());
-    }
-
-    /**
-     * CDI Dispose method for {@link #topicAdminClient()}. Shouldn't be called directly
-     */
-    public void shutdownTopicAdminClient(@Disposes TopicAdminClient topicAdminClient) {
-        topicAdminClient.close();
     }
 
     /**
